+++ packages/layer-composer/src/generators/context/context-layers.ts
@@ -2,6 +2,8 @@ import type { FillLayer, Layer, LineLayer } from '@globalfishingwatch/mapbox-gl'
 import { Group } from '../../types'
 import { ContextLayerType } from '../types'
 
+export const HIGHLIGHT_SUFIX = '_highlight'
+
 const settledBoundaries = [
   '200 NM',
   'Treaty',
@@ -71,7 +73,7 @@ const getDefaultContextLayersById = (id: string, color: string): (LineLayer | Fi
       ...getDefaultContextLine(color),
     } as LineLayer,
     {
-      id: `${id}-highlight`,
+      id: `${id}${HIGHLIGHT_SUFIX}`,
       ...getDefaultContextHighlight(),
     } as LineLayer,
   ]
@@ -93,7 +95,7 @@ const CONTEXT_LAYERS: Record<ContextLayerType, Layer[]> = {
       ...getDefaultContextInteraction(),
     } as FillLayer,
     {
-      id: 'eez_highlight_',
+      id: `eez${HIGHLIGHT_SUFIX}_`,
       ...getDefaultContextHighlight(),
     } as LineLayer,
   ],