+++ applications/amathea/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @globalfishingwatchapp/amathea
 
+## 0.14.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @globalfishingwatch/layer-composer@4.14.5
+  - @globalfishingwatch/react-hooks@4.8.3
+  - @globalfishingwatch/ui-components@2.13.6
+
 ## 0.14.3
 
 ### Patch Changes