+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @globalfishingwatch/layer-composer@4.14.5
+
 ## 1.1.13
 
 ### Patch Changes