+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @globalfishingwatch/react-hooks
 
+## 4.8.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @globalfishingwatch/layer-composer@4.14.5
+
 ## 4.8.2
 
 ### Patch Changes